refactor(config): move Slack and Signal schemas to plugins
