+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -22,7 +22,7 @@ ad60ccc4fe9084d47f0477e02d9296bacad32f26d7456e2a84be8d25a53a25c2  module/boolean
 6e3b8ccee738f8eddde5d813b6e204d6f7c8541fda4cb279a3be3ead27f47e7d  module/channel-actions
 00d326fdff396dfcc657d1fd16a4015b81d60ae8bf7dad09ac1010b376814de6  module/channel-config-helpers
 16dc9d32e8ca3ef78fc63e0fc4b20e6b943633e9572de1a49d24a880b6ffc66c  module/channel-config-primitives
-5da2caccf30780a4cf86ac710e2ca42cc9576cb1c49291c0929889fec0cc2257  module/channel-config-schema
+f80effbeef28fefc635e8cdc85fcc9d9d453325cd0478b7db5914abe48582bd4  module/channel-config-schema
 2dd98659d9600e755f09ef00dd91c36692b8562ed3700461937e94f6cd1e640a  module/channel-contract
 a6bc51efd3ce23e65432c9a6790670e2a0acb96ae7263fedbaed52ea353228be  module/channel-core
 fbf353eb38ae68d8ded3f2a60b432c7bb2c245d2ec7e7c9f53c6da19a0db0938  module/channel-entry-contract
