+++ extensions/signal/config-api.ts
@@ -1,3 +1,3 @@
 // Signal API module exposes the plugin public contract.
 export { buildChannelConfigSchema } from "openclaw/plugin-sdk/channel-config-schema";
-export { SignalConfigSchema } from "openclaw/plugin-sdk/bundled-channel-config-schema";
+export { SignalConfigSchema } from "./src/config-schema.js";
