+++ extensions/slack/config-api.ts
@@ -1,3 +1,3 @@
 // Slack API module exposes the plugin public contract.
 export { buildChannelConfigSchema } from "openclaw/plugin-sdk/channel-config-schema";
-export { SlackConfigSchema } from "openclaw/plugin-sdk/bundled-channel-config-schema";
+export { SlackConfigSchema } from "./src/config-schema.js";
