+++ src/plugin-sdk/channel-config-schema.ts
@@ -15,16 +15,36 @@ export {
   buildNestedDmConfigSchema,
 } from "../channels/plugins/config-schema.js";
 export {
+  BlockStreamingChunkSchema,
   BlockStreamingCoalesceSchema,
+  ChannelDeliveryStreamingConfigSchema,
+  ChannelStreamingBlockSchema,
   ContextVisibilityModeSchema,
   DmConfigSchema,
   DmPolicySchema,
+  ExecutableTokenSchema,
   GroupPolicySchema,
   MarkdownConfigSchema,
   MentionPatternsPolicySchema,
+  ProviderCommandsSchema,
+  ReplyToModeSchema,
   ReplyRuntimeConfigSchemaShape,
+  TextChunkModeSchema,
   requireAllowlistAllowFrom,
   requireOpenAllowFrom,
 } from "../config/zod-schema.core.js";
+export {
+  buildChannelAllowBotsSchema,
+  buildChannelExecApprovalsSchema,
+  buildChannelReactionShape,
+  buildCommonChannelAccountShape,
+  ChannelBotLoopProtectionSchema,
+  ChannelDangerouslyAllowNameMatchingSchema,
+  ChannelPreviewStreamingConfigSchema,
+  ChannelSendReadReceiptsSchema,
+  ChannelStreamingProgressSchema,
+  ChannelStreamingPreviewSchema,
+  UnifiedStreamingModeSchema,
+} from "../config/zod-schema.channel-messaging-common.js";
 export { ChannelImplicitMentionsSchema } from "../config/zod-schema.implicit-mentions.js";
 export { ToolPolicySchema } from "../config/zod-schema.agent-runtime.js";
