+++ src/plugins/contracts/config-footprint-guardrails.test.ts
@@ -139,14 +139,19 @@ describe("config footprint guardrails", () => {
     }
   });
 
-  it("keeps canonical nested streaming paths in the public core channel schema", () => {
+  it("keeps canonical nested streaming paths in channel-owned schemas", () => {
     const source = readSource("src/config/zod-schema.providers-core.ts");
+    const slackSource = readSource("extensions/slack/src/config-schema.ts");
 
     expect(source).toContain("streaming: ChannelPreviewStreamingConfigSchema.optional(),");
-    expect(source).toContain("streaming: SlackStreamingConfigSchema.optional(),");
-    expect(source).not.toContain('streamMode: z.enum(["replace", "status_final", "append"])');
-    expect(source).not.toContain("draftChunk:");
-    expect(source).not.toContain("nativeStreaming:");
+    expect(slackSource).toContain("streaming: SlackStreamingConfigSchema.optional(),");
+    for (const schemaSource of [source, slackSource]) {
+      expect(schemaSource).not.toContain(
+        'streamMode: z.enum(["replace", "status_final", "append"])',
+      );
+      expect(schemaSource).not.toContain("draftChunk:");
+      expect(schemaSource).not.toContain("nativeStreaming:");
+    }
   });
 
   it("keeps Matrix setup input canonical-first after plugin ownership", () => {
@@ -205,8 +210,6 @@ describe("config footprint guardrails", () => {
       "GoogleChatConfigSchema",
       "IMessageConfigSchema",
       "MSTeamsConfigSchema",
-      "SignalConfigSchema",
-      "SlackConfigSchema",
       "TelegramConfigSchema",
       "WhatsAppConfigSchema",
     ]);
